- Sales module finished.

+++ service-domain/src/main/java/com/programacion/ecommerce/entities/ProductEntity.java
@@ -31,7 +31,7 @@ public class ProductEntity extends BaseEntity {
 
     @NotNull
     @Column(name = "price", precision = 10, scale = 2, nullable = false)
-    private Integer price;
+    private BigDecimal price;
 
     @NotNull
     @Enumerated(EnumType.STRING)
